refactor: controller와 view 분리
