+++ src/js/controller/handlePurchaseResultToggle.js
@@ -1,19 +1,12 @@
-import { hideElement, showElement } from '../utils/setViewProperty.js';
+import {
+  renderPurchaseResultSectionColAlign,
+  renderPurchaseResultSectionRowAlign,
+} from '../view/viewPurchaseResultSection.js';
 
 export const handlePurchaseResultToggle = ({ target }) => {
-  const $purchaseResultSectionRowAlign = document.querySelector(
-    '#purchase-result-section__row-align',
-  );
-  const $purchaseResultSectionColAlign = document.querySelector(
-    '#purchase-result-section__col-align',
-  );
-
   if (target.checked) {
-    showElement($purchaseResultSectionColAlign);
-    hideElement($purchaseResultSectionRowAlign);
-    return;
+    return renderPurchaseResultSectionColAlign();
   }
 
-  showElement($purchaseResultSectionRowAlign);
-  hideElement($purchaseResultSectionColAlign);
+  renderPurchaseResultSectionRowAlign();
 };
