+++ src/js/view/setViewProperty.js
